+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Comerce
 
 - `/commerce/trackPurchase` endpoint added as `Iterable::Commerce#track_purchase`
+- `/commerce/updateCart` endpoint added as `Iterable::Commerce#update_cart`
 
 ## Device